+++ gulp/cache.js
@@ -1,26 +1,16 @@
 'use strict';
 var path         = require('path');
 var gulp         = require('gulp');
-var replace       = require('replace');
-var paths        = require('./config');
-var replaceFiles = ['./www/js/app.js'];
+var replace      = require('replace');
+var paths        = require('./config').paths;
+var replaceFiles = [paths.src + '/js/app.js'];
 
 // Cache Modules
 // ADD
 gulp.task('cacheapp:add', function () {
     return replace({
-        regex      : "//'cacheapp'",
-        replacement: "'cacheapp'",
-        paths      : replaceFiles,
-        recursive  : false,
-        silent     : false
-    });
-});
-
-gulp.task('cachemodule:add', function () {
-    return replace({
-        regex      : "//'cachemodule'",
-        replacement: "'cachemodule'",
+        regex      : "//'app.cache'",
+        replacement: "'app.cache'",
         paths      : replaceFiles,
         recursive  : false,
         silent     : false
@@ -30,17 +20,8 @@ gulp.task('cachemodule:add', function () {
 // REMOVE
 gulp.task('cacheapp:remove', function () {
     return replace({
-        regex      : "'cacheapp'",
-        replacement: "//'cacheapp'",
-        paths      : replaceFiles,
-        recursive  : false,
-        silent     : false
-    });
-});
-gulp.task('cachemodule:remove', function () {
-    return replace({
-        regex      : "'cachemodule'",
-        replacement: "//'cachemodule'",
+        regex      : "'app.cache'",
+        replacement: "//'app.cache'",
         paths      : replaceFiles,
         recursive  : false,
         silent     : false