+++ gulp/jshint.js
@@ -1,13 +1,13 @@
 'use strict';
 var path    = require('path');
 var gulp    = require('gulp');
-var paths   = require('./config');
 var jshint  = require('gulp-jshint');
 var stylish = require('jshint-stylish');
+var paths   = require('./config').paths;
 
 // Lint
 gulp.task('jshint', function () {
-    return gulp.src(paths.src.js)
+    return gulp.src(paths.src + '/js/**/*.js')
         .pipe(jshint())
         .pipe(jshint.reporter(stylish));
 });