+++ gulp/templates.js
@@ -4,36 +4,23 @@ var gulp          = require('gulp');
 var minifyHTML    = require('gulp-minify-html');
 var templateCache = require('gulp-angular-templatecache');
 var iife          = require("gulp-iife");
-var paths         = require('./config');
+var paths         = require('./config').paths;
 
 // Templates
 gulp.task('template:app', function () {
-    gulp.src(['./www/app/**/*.html'])
+    gulp.src([paths.src + '/js/**/*.html'])
         .pipe(minifyHTML({quotes: true}))
         .pipe(templateCache({
-            module    : 'cacheapp',
-            filename: 'cacheapp.js',
+            module    : 'app.cache',
+            filename: 'app.cache.js',
             root    : 'app',
             standalone: true
         }))
         .pipe(iife())
-        .pipe(gulp.dest('./www/js/'));
+        .pipe(gulp.dest(paths.src + '/js/'));
 });
 
-gulp.task('template:module', function () {
-    gulp.src(['./www/module/**/*.html'])
-        .pipe(minifyHTML({quotes: true}))
-        .pipe(templateCache({
-            module    : 'cachemodule',
-            filename: 'cachemodule.js',
-            root    : 'module',
-            standalone: true
-        }))
-        .pipe(iife())
-        .pipe(gulp.dest('./www/js/'));
-});
 
 gulp.task('templates', [
-    'template:app',
-    'template:module'
+    'template:app'
 ]);