+++ www/component/ion-crop/ion-crop.js
@@ -1,6 +1,0 @@
-(function (window, angular, undefined) {
-  'use strict';
-  'use strict';
-  angular.module('ion-crop', []);
-
-})(window, window.angular);