+++ www/js/app.js
@@ -5,31 +5,29 @@
       'ionic',
       'ionic.service.core',
       'ionic.service.analytics',
-      //'cacheapp',
-      //'cachemodule',
       'ionic-cache-src',
       'formlyIonic',
       'pascalprecht.translate',
       'angularMoment',
       'ionic.components',
       'ngFacebook',
-      'translate.app',
-      'translate.form',
       'angular-cache',
       'ngCordova',
+      'translate.form',
       'gettext',
-      'module.core',
-      'module.user',
-      'module.gallery'
+      //'app.cache',
+      'app.translate',
+      'app.photogram',
     ])
-
-
-  .run(runIonic)
+    .run(runIonic)
     .run(runLanguage)
-    .config(config);
+    .run(runFacebook)
+    .config(configFacebook)
+    .config(configIonic);
+
 
   function runIonic($ionicPlatform, $cacheSrc, AppConfig, $ionicAnalytics, $cordovaStatusbar, $timeout,
-    $cordovaSplashscreen, GallerySetting, User) {
+    $cordovaSplashscreen, PhotogramSetting, User) {
 
     $cacheSrc.color = AppConfig.color;
     $cacheSrc.bgcolor = '#ddd';
@@ -38,7 +36,7 @@
     //$cacheSrc.interval = 5000;
 
     User.init();
-    GallerySetting.init();
+    PhotogramSetting.init();
 
     $ionicPlatform.ready(function () {
 
@@ -92,12 +90,11 @@
     console.info(LangVar, userLangVar);
   }
 
-  function config($ionicConfigProvider) {
+  function configIonic($ionicConfigProvider) {
     //$ionicConfigProvider.platform.ios.backButton.previousTitleText(' ').icon('ion-ios-arrow-left');
     //$ionicConfigProvider.platform.android.backButton.previousTitleText(' ').icon('ion-ios-arrow-left');
     //$ionicConfigProvider.views.swipeBackEnabled (true);
-    $ionicConfigProvider.backButton.text(' ')
-      .icon('ion-ios-arrow-left');
+    $ionicConfigProvider.backButton.text(' ').icon('ion-ios-arrow-left');
     //$ionicConfigProvider.backButton.previousTitleText (false).text ('Voltar').icon ('ion-ios-arrow-left');
     //$ionicConfigProvider.views.transition ('platform');
     //$ionicConfigProvider.navBar.alignTitle ('platform');
@@ -108,6 +105,37 @@
     // Android Native Scroll
     var jsScrolling = (ionic.Platform.isAndroid()) ? false : true;
     $ionicConfigProvider.scrolling.jsScrolling(jsScrolling);
+  }
+
+  // Facebook
+
+  function runFacebook() {
+    var ionic = window.ionic;
+    if (!(ionic.Platform.isIOS() || ionic.Platform.isAndroid())) {
+      var LangVar = window.navigator.language || window.navigator.userLanguage;
+      var userLangVar = LangVar.substring(0, 2) + '_' + LangVar.substring(3, 5).toUpperCase();
+      var document = window.document;
+      // If we've already installed the SDK, we're done
+      if (document.getElementById('facebook-jssdk')) return;
+
+      // Get the first script element, which we'll use to find the parent node
+      var firstScriptElement = document.getElementsByTagName('script')[0];
+
+      // Create a new script element and set its id
+      var facebookJS = document.createElement('script');
+      facebookJS.id = 'facebook-jssdk';
+
+      // Set the new script's source to the source of the Facebook JS SDK
+      facebookJS.src = 'http://connect.facebook.net/' + userLangVar + '/all.js';
+
+      // Insert the Facebook JS SDK into the DOM
+      firstScriptElement.parentNode.insertBefore(facebookJS, firstScriptElement);
+    }
+  }
+
+  function configFacebook($facebookProvider, AppConfig) {
+    $facebookProvider.setAppId(AppConfig.facebook);
+    $facebookProvider.setPermissions('id,name,email,user_likes,bio');
   }