+++ www/js/components/components.module.js
@@ -1,0 +1,12 @@
+(function (window, angular, undefined) {
+    'use strict';
+    angular.module('ionic.components', [
+        'ionic',
+        'ionic-loading',
+        //'ion-affix',
+        'ion-location',
+        'ion-photo',
+        'ion-crop',
+        'ion-caman'
+    ]);
+})(window, window.angular);