+++ www/module/gallery/module.gallery.js
@@ -1,4 +1,0 @@
-(function (window, angular, undefined) {
-  'use strict';
-  angular.module('module.gallery', []);
-})(window, window.angular);